Move logger into a shared core package.
