+++ apps/task-master/package.json
@@ -7,6 +7,7 @@
     "dev": "env-cmd -f .env.development ts-node ./src/main.ts"
   },
   "dependencies": {
+    "@repo/core": "*",
     "@repo/database": "*",
     "env-cmd": "^10.1.0"
   },
