+++ apps/web/package.json
@@ -20,6 +20,7 @@
     "@next-auth/prisma-adapter": "^1.0.7",
     "@react-pdf-viewer/core": "3.12.0",
     "@react-pdf-viewer/page-navigation": "^3.12.0",
+    "@repo/core": "*",
     "@repo/database": "*",
     "@svgr/webpack": "8.1.0",
     "@types/lodash": "^4.14.197",
@@ -61,9 +62,7 @@
     "swr": "^2.2.4",
     "tailwindcss": "^3.3.3",
     "twind": "0.16.19",
-    "ua-parser-js": "^1.0.36",
-    "winston": "^3.10.0",
-    "winston-daily-rotate-file": "^4.7.1"
+    "ua-parser-js": "^1.0.36"
   },
   "devDependencies": {
     "@repo/typescript-config": "*",
