+++ apps/web/pages/api/organizations/[orgIdOrSlug]/index.ts
@@ -14,7 +14,6 @@ import { OrganizationResponse } from "lib/types/api/organization.response";
 import { OrganizationService } from "lib/api/services/organization-service";
 import { OrgMembershipService } from "lib/api/services/org-membership-service";
 import { OrganizationUpdateRequest } from "lib/types/api/organization-update.request";
-import { ModelType } from "lib/types/core/model-type";
 
 const organizationService = new OrganizationService();
 const orgMembershipService = new OrgMembershipService();
